[Justice Counts] Trying a couple things to improve performance (Recidiviz/recidiviz-data#22256)
## Description of the change

This PR does 3 things, 2 of which are perf related, and 1 that I am just
sneaking in:
1) Reduce our number of gunicorn workers. We currently have something
like 11 workers, which is a) using a lot of memory, and b) resulting in
11 different database connection pools (more info
[here](https://davidcaron.dev/sqlalchemy-multiple-threads-and-processes/)
-- essentially each worker will get its own pool, containing a single
connection to the database), which ends up meaning that many requests to
our app have to establish a new connection to the DB, rather than
re-using an existing one. Establishing a connection to the DB takes
~400ms which really slows things down!
2) Increase the `pool_recycle` parameter from 10mins to 1hr. This means
that database connections will expire less frequently. Every time a
connection expires, we have to create a new one, which -- as above --
takes a long time! As per [this
PR](https://github.com/Recidiviz/recidiviz-data/pull/20379), it's
possible this will start resulting in `sqlalchemy.exc.OperationalError`
errors -- I will keep an eye on this and revert if so.
3) Update the deploy_for_playtesting script so we can deploy Agency
Dashboards to staging.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Contributes to Recidiviz/recidiviz-data#22101

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0cad122c8c4180492953c30d1083494a46453108