[US_CA] Parse 'CS P36 Specialized Caseload' `StateSupervisionLevel` (Recidiviz/recidiviz-data#22334)
## Description of the change

It seems like `CS P36 Specialized Caseload` was either a
`SupervisionLevel` value in the `supervision_period` ingest view results
that slipped in the cracks until now, or one that just started
appearing. Ingest is crashing because it isn't mapped to a
`StateSupervisionLevel` in the mapping.

I added it to be ignored here, because it didn't obviously fit into any
of the other categories with my limited context. @not-a-doctor-stromberg
let me know if you think this should go somewhere else!

I also ran the `run_mappings_against_ingest_view_results` script on this
view/mapping to make sure no other enum values were missing from the
mapping. It parsed successfully, so there shouldn't be any more missing
values here as of now.

Since this is currently causing an extract and merge error and hasn't
caused any bad data to appear in state dataset or elsewhere, the only
thing needed to fully implement this change is for this PR to be merged
into main and deployed. I will CP this commit into the production deploy
happening tomorrow (7/18) once it has been merged into main.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22330

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6f3e70bd12bf879406cb280f3b734ea45d45b3fc