Match CA supervision officers by name in lieu of badge numbers (Recidiviz/recidiviz-data#22217)
## Description of the change

Because we may not have supervisor badge numbers for clients soon
enough, this PR looks up badge numbers based on the parole agent's name
and unit.

This seems to work pretty well, but not perfectly: out of 42k client
records, 40k got officer_ids using the old method (the archived import)
and 35k get officer_ids using this method.

Three questions I have:
* ~20% of rows in PersonParole_latest have a BadgeNumber. Are those
trustworthy? This PR treats those as the best source of truth.
* Some (not many) unit/name combos in AgentParole_latest map to multiple
BadgeNumbers. I picked the MIN() because it seems to cause the fewest
discrepancies with the PersonParole badges we do have. Can you think of
a more principled approach here?
* Is it worth it to go full belt-and-suspenders and use the archived
import for rows that still need a match?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 40b65e5521510a2bea6fe2e86beabc485456ec95