[Bugfix] Update Airflow entrypoints to disallow unknown args (Recidiviz/recidiviz-data#22291)
## Description of the change

No reason to use `parse_known_args` when we don't expect there to be
unknown args. Updates Airflow entrypoint scripts to use `parse_args`
instead. There are a bunch of other places where we use
`parse_known_args` which I am not tackling here, but figured this was a
good place to start establishing this pattern.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Motivated by
[this](https://recidiviz.slack.com/archives/C028X32LRH7/p1689360828010509)
slack thread, though this fix ended up being not relevant to the issue.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ce22d5979ac7a9144396b1b137c100a2f0496026