[Justice Counts] Get Bulk Upload errors/warnings json (Recidiviz/recidiviz-data#21959)
## Description of the change

In order to supply a stand-alone errors/warnings and metric review page,
we need to retrieve the stored json that contains the errors/warnings
and review data from the GCP bucket. This PR adds a GET endpoint
`/<spreadsheet_id>/bulk-upload-json`, which will supply the frontend
with the json produced by a given spreadsheet.

Demonstration of this code in python shell:

https://github.com/Recidiviz/recidiviz-data/assets/82831800/3c807498-7669-4804-8f70-f58b0be6fb5e

Accompanying FE PRs:
https://github.com/Recidiviz/justice-counts/pull/758 and
https://github.com/Recidiviz/justice-counts/pull/767

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#21940

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: 6db5317e399c389b3c0121ae2daa1f36ace736ad