Add `StateIncarcerationPeriodHousingUnitCategory` & granularity to `StateIncarcerationPeriodHousingUnitType` (Recidiviz/recidiviz-data#22220)
## Description of the change

Implements [this schema change
request](https://paper.dropbox.com/doc/APPROVED-Schema-change-request-Additional-Enums-for-Solitary-Confinement--B790jh8TNQrMIR75eFXdmL4VAg-1LPdg4IRpdJiZktzHziG6).

Adds a `StateIncarcerationPeriodHousingUnitCategory` field to
`StateIncarcerationPeriod` to serve as a parent field for
`StateIncarcerationPeriodHousingUnitType`. This new field can have the
following values:
- `StateIncarcerationPeriodHousingUnitCategory.SOLITARY_CONFINEMENT`:
"This person has been assigned to a solitary confinement unit for an
indeterminate amount of time."
- `StateIncarcerationPeriodHousingUnitCategory.GENERAL`: "This person is
in a non-specialty housing unit. Incarceration periods will generally be
assigned this value by default."

Adds granularity to the existing
`StateIncarcerationPeriodHousingUnitType` field of
`StateIncarcerationPeriod`. This enum can now have the following values:
-
`StateIncarcerationPeriodHousingUnitType.TEMPORARY_SOLITARY_CONFINEMENT`:
"This is the placement of a person in restrictive housing that can occur
for a wide range of institutional needs and likely has another bed saved
for them elsewhere in the facility. For example, it can be an interim
status for people pending their transfer to another institution or
awaiting a judicial proceeding, to facilitate a criminal investigation,
or when limited bed space in an institution necessitates the use of an
otherwise empty segregation cell."
-
`StateIncarcerationPeriodHousingUnitType.DISCIPLINARY_SOLITARY_CONFINEMENT`:
"This person is placed in restrictive housing as a form of punishment."
-
`StateIncarcerationPeriodHousingUnitType.ADMINISTRATIVE_SOLITARY_CONFINEMENT`:
"This person is placed in restrictive housing for managerial purposes,
including as a response to a person who demonstrates a chronic inability
to adjust to the general population, or when authorities believe an
inmate's presence in the general population may cause a serious
disruption to the orderly operation of the institution."
- `StateIncarcerationPeriodHousingUnitType.PROTECTIVE_CUSTODY`: "This
person is placed in restrictive housing to separate vulnerable people
from the general population due to personal physical safety concerns."
- `StateIncarcerationPeriodHousingUnitType.OTHER_SOLITARY_CONFINEMENT`:
"This person is placed in a restrictive housing unit that is not
temporary but does not fit into the other categories."
-
`StateIncarcerationPeriodHousingUnitType.MENTAL_HEALTH_SOLITARY_CONFINEMENT`:
"This person is placed in restrictive housing because of their mental
health status."
- `StateIncarcerationPeriodHousingUnitType.HOSPITAL`: "This person is
placed in a hospital unit for medical attention."
- `StateIncarcerationPeriodHousingUnitType.GENERAL`: "This person is in
a non-specialty housing unit. Incarceration periods will generally be
assigned this value by default."
-
`StateIncarcerationPeriodHousingUnitType.PERMANENT_SOLITARY_CONFINEMENT`:
Will maintain this legacy enum value as an option until it is deprecated
by Recidiviz/recidiviz-data#22252

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes  Recidiviz/recidiviz-data#21777

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b7cff074d4d3d32c1e7b351dc8866b72b6c8504e