Revise validation to handle null level 2 supervision levels (Recidiviz/recidiviz-data#22402)
## Description of the change

The `location_ids_to_names_unique_ids.py` validation is failing for
[US_IX](https://recidiviz-staging.ue.r.appspot.com/admin/validation_metadata/status/details/location_ids_to_names_unique_ids?stateCode=US_IX)
because the `CONCAT` in the validation returns NULL if
`level_2_supervision_location_external_id` is NULL (which is the case
for a few entries in location metadata). This PR revises the validation
to account for NULLs with a `COALESCE`.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9857530a2eb6da39dadc65d3727accc8f90b487e