[US_IX] Add `Release to Parole` as erroneous `StateIncarcerationPeriodAdmissionReason` enum value (Recidiviz/recidiviz-data#22413)
## Description of the change

Ingest is crashing because `Release to Parole` started appearing in the
`Start_TransferReasonDesc` field, which we map to
`StateIncarcerationPeriodAdmissionReason`. This change maps that value
to `INTERNAL_UNKNOWN` and adds it to Recidiviz/recidiviz-data#21085 be investigated later.

Since this is failing in staging and prod, this change will need to be
cherry-picked into prod for ingest to run in prod.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22412

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d1caf057593ac07a2e9d1d103daf3841fe6e2b45