[US_IX] Fix premature state staff role period end dates (Recidiviz/recidiviz-data#22392)
## Description of the change

The `state_staff_role_location_periods` view was prematurely ending role
periods and location periods. This is because it was delusionally trying
to account for the scenario where employees get deleted out of the raw
Atlas employee table, when that's not a scenario we would ever observe
because we get two-week diffs. So instead it was accidentally ending
periods for employees that simply haven't had their record updated in
the last two weeks.

This PR removes the filter that guards for the deletion scenario which
allows us to also simplify the view a bit. It also adds an
`always_historical` flag for the raw data config of the legacy employee
table (which is not used in this view).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bbf9f93ff32680ba2c1f61a6b18bebdc46e46d83