[US_ND] Update incarceration sentences view (Recidiviz/recidiviz-data#22384)
## Description of the change

When I refactored the views that hydrate incarceration_sentence and
state_charge in ND, I noticed that all but one of the relevant raw data
tables format the `OFFENDER_BOOK_ID` field like `123,456.00` when the ID
is actually `123456`. In the combined `elite_incarceration_sentences`
view, I worked around this by trimming `OFFENDER_BOOK_ID` values from
those files using this command:
`TRIM(REPLACE(OFFENDER_BOOK_ID, ',',''), '.00')`

I didn't know that the `TRIM` function in SQL removes _all_ instances of
_any_ character in the string argument. Because of this, unique
`OFFENDER_BOOK_ID` values like `110,000.00` and `1,100.00` were both
being simplified to `11`, when the intent was for them to be trimmed to
`110000` and `1100`.

This change replaces the `TRIM` with a second call to `REPLACE` in each
of these instances, so the general structure is
`REPLACE(REPLACE(OFFENDER_BOOK_ID, ',',''), '.00', '')`. This is what I
was attempting to do to begin with!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22326

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3cb3d6a9bd66fd059c7298e7917d5c13789a4c67