[US_PA] Update `us_pa_location_metadata` to include districts (Recidiviz/recidiviz-data#22417)
## Description of the change

Since we are including District Directors and Deputy District Directors
in PA outliers views and use `district_id` as their location ID, there
would ideally be rows in `location_metadata` that correspond to
districts rather than only the more granular supervision units. This
would allow us to easily use the `location_metadata` reference table for
upper management as well as unit supervisors. Looking ahead, we also
want to be able to do this for regional managers.

This change appends the 10 district IDs and 3 region IDs in PA to the
`location_external_id` column in `location_metadata`. In the JSON blob
that pulls every level of location information into `location_metadata`,
the fields corresponding to supervision unit will be blank for rows that
represent supervision districts
(`level_1_supervision_location_external_id` and
`level_1_supervision_location_name` in the
`location_metadata.location_metadata` json blob field), and the
supervision unit and district fields will be blank in rows that
represent supervision regions.

This change also pads district IDs with zeros in the
`state_staff_location_period` ingest view so that the IDs will match the
reference table.

If this change is approved, I will run the updated query in the raw data
config here and upload the new
`RECIDIVIZ_REFERENCE_supervision_location_ids` file to be ingested,
which will update `us_pa_location_metadata`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#22065

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 49f81ed9d8e2d844446b2ef7a3829057b9530fed